Cleaning up lots of unneeded code and files
