+++ src/main/webapp/app/controls/course/course-card/course-card.component.css
@@ -1,8 +1,5 @@
 .example-card {
   margin: 10px;
- 
-  /*width: 400px;*/
-  /*height: 100%;*/
 }
 
 .example-header-image {
