+++ src/main/webapp/app/controls/course/course-card/course-card.component.html
@@ -7,9 +7,7 @@
   </md-card-header>
   <img md-card-image src="http://image.prntscr.com/image/8d081d23bd584342ba119fb562510f37.png">
   <md-card-content>
-    <p>
-     {{course.description}}
-    </p>
+    <p>{{course.description}}</p>
   </md-card-content>
   <md-card-actions>
     <button md-button *ngIf="!studentView">LIKE</button>
