+++ src/main/webapp/app/controls/course/course-list/course-list.component.html
@@ -1,3 +1,3 @@
 <div fxLayout="row" fxLayoutWrap="wrap" class="container" fxLayoutAlign="center center">
-  <app-course-card fxFlex.xs="100" fxFlex.sm="50" fxFlex.md="28" fxFlex.lg="20" fxFlex.xl="15" *ngFor="let course of courses" [course]="course" [studentView]="studentView"></app-course-card>
+  <app-course-card fxFlex.xs="100" fxFlex.sm="50" fxFlex.md="28" fxFlex.lg="20" fxFlex.xl="15"*ngFor="let course of courses" [course]="course" [studentView]="studentView"></app-course-card>
 </div>
