+++ src/main/webapp/app/controls/course/course-list/course-list.component.ts
@@ -1,17 +1,14 @@
-import { Course } from './../../../models/course';
-import { Component, OnInit, Input } from '@angular/core';
+import {Component, Input} from '@angular/core';
+
+import {Course} from '../../../models/course';
 
 @Component({
   selector: 'app-course-list',
   templateUrl: './course-list.component.html',
   styleUrls: ['./course-list.component.css']
 })
-export class CourseListComponent implements OnInit {
+export class CourseListComponent {
 
-  constructor() { }
   @Input() courses: Course[] = [];
   @Input() studentView: boolean = false;
-  ngOnInit() {
-  }
-
 }
