--- src/main/webapp/app/controls/course/course-milestone-list/course-milestone-list.component.css
@@ -1,15 +1,0 @@
-.milestone-completed {
-  background-color: rgba(0,211,167,1);
-  border-radius: 50%;
-  display: inline-block;
-  width: 30px;
-  height: 30px;
-  color: white;
-  text-align: center;
-  margin-right: 10px;
-}
-
-.milestone-points {
-  text-align: right;
-  color: rgba(204,204,204,1);
-}
