--- src/main/webapp/app/controls/course/course-milestone-list/course-milestone-list.component.ts
@@ -1,16 +1,0 @@
-import { Milestone } from './../../../models/milestone';
-import { Component, OnInit, Input } from '@angular/core';
-
-@Component({
-  selector: 'app-course-milestone-list',
-  templateUrl: './course-milestone-list.component.html',
-  styleUrls: ['./course-milestone-list.component.css']
-})
-export class CourseMilestoneListComponent implements OnInit {
-
-  constructor() { }
-  @Input() milestones: Milestone[] = [];
-  ngOnInit() {
-  }
-
-}
