+++ src/main/webapp/app/controls/layout/parallax-header/parallax-header.component.ts
@@ -1,15 +1,8 @@
-import { Component, OnInit } from '@angular/core';
+import {Component} from '@angular/core';
 
 @Component({
   selector: 'app-parallax-header',
   templateUrl: './parallax-header.component.html',
   styleUrls: ['./parallax-header.component.scss']
 })
-export class ParallaxHeaderComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
+export class ParallaxHeaderComponent {}
