--- src/main/webapp/app/controls/portfolio/portfolio-card/portfolio-card.component.spec.ts
@@ -1,25 +1,0 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { PortfolioCardComponent } from './portfolio-card.component';
-
-describe('PortfolioCardComponent', () => {
-  let component: PortfolioCardComponent;
-  let fixture: ComponentFixture<PortfolioCardComponent>;
-
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [ PortfolioCardComponent ]
-    })
-    .compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(PortfolioCardComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
