+++ src/main/webapp/app/controls/portfolio/portfolio-card/portfolio-card.component.ts
@@ -1,20 +1,15 @@
-import { Component, OnInit, Input } from '@angular/core';
-import { PortfolioItem } from '../../../models/portfolio-item'
+import {Component, Input} from '@angular/core';
+
+import {PortfolioItem} from '../../../models/portfolio-item'
 
 @Component({
   selector: 'app-portfolio-card',
   templateUrl: './portfolio-card.component.html',
   styleUrls: ['./portfolio-card.component.css']
 })
-export class PortfolioCardComponent implements OnInit {
+export class PortfolioCardComponent {
 
   @Input() portfolioItem: PortfolioItem;
-
-  constructor() {
-  }
-
-  ngOnInit() {
-  }
 
   share() {
     alert(this.portfolioItem.filename);
