--- src/main/webapp/app/controls/portfolio/portfolio-list/portfolio-list.component.spec.ts
@@ -1,25 +1,0 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { PortfolioListComponent } from './portfolio-list.component';
-
-describe('PortfolioListComponent', () => {
-  let component: PortfolioListComponent;
-  let fixture: ComponentFixture<PortfolioListComponent>;
-
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [ PortfolioListComponent ]
-    })
-    .compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(PortfolioListComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should be created', () => {
-    expect(component).toBeTruthy();
-  });
-});
