+++ src/main/webapp/app/controls/portfolio/portfolio-list/portfolio-list.component.ts
@@ -1,17 +1,13 @@
-import { Component, OnInit, Input } from '@angular/core';
-import { PortfolioItem } from './../../../models/portfolio-item';
+import {Component, Input} from '@angular/core';
+
+import {PortfolioItem} from '../../../models/portfolio-item';
 
 @Component({
   selector: 'app-portfolio-list',
   templateUrl: './portfolio-list.component.html',
   styleUrls: ['./portfolio-list.component.css']
 })
-export class PortfolioListComponent implements OnInit {
+export class PortfolioListComponent {
 
-  constructor() { }
   @Input() portfolios: PortfolioItem[] = [];
-  ngOnInit() {
-
-  }
-
 }
