+++ src/main/webapp/app/models/account.model.ts
@@ -1,0 +1,11 @@
+export class AccountModel {
+  constructor(
+    public activated: boolean,
+    public authorities: string[],
+    public email: string,
+    public firstName: string,
+    public lastName: string,
+    public login: string,
+    public imageUrl: string
+  ) {}
+}
