+++ src/main/webapp/app/models/course.ts
@@ -1,4 +1,4 @@
-import {User} from "../shared/user/user.model";
+import {User} from "./user.model";
 
 export interface Course {
   name?: string;
