+++ src/main/webapp/app/pages/access-denied-page/access-denied-page.component.ts
@@ -1,15 +1,8 @@
-import { Component, OnInit } from '@angular/core';
+import {Component} from '@angular/core';
 
 @Component({
   selector: 'app-access-denied-page',
   templateUrl: './access-denied-page.component.html',
   styleUrls: ['./access-denied-page.component.css']
 })
-export class AccessDeniedPageComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
+export class AccessDeniedPageComponent {}
