+++ src/main/webapp/app/services/login.service.ts
@@ -1,8 +1,8 @@
 import { Router } from '@angular/router';
 import { Injectable } from '@angular/core';
 
-import { Principal } from '../auth/principal.service';
-import { AuthServerProvider } from '../auth/auth-oauth2.service';
+import { Principal } from '../shared/auth/principal.service';
+import { AuthServerProvider } from '../shared/auth/auth-oauth2.service';
 
 @Injectable()
 export class LoginService {
