+++ src/main/webapp/app/shared/index.ts
@@ -5,10 +5,9 @@ export * from './auth/auth-oauth2.service';
 export * from './auth/auth.service';
 export * from './auth/principal.service';
 export * from './auth/has-any-authority.directive';
-export * from './login/login.service';
-export * from './user/account.model';
-export * from './user/user.model';
-export * from './user/user.service';
+export * from '../services/login.service';
+export * from '../models/account.model';
+export * from '../models/user.model';
 export * from './shared-libs.module';
 export * from './shared-common.module';
 export * from './shared.module';
