--- src/main/webapp/app/shared/user/account.model.ts
@@ -1,11 +1,0 @@
-export class AccountModel {
-    constructor(
-        public activated: boolean,
-        public authorities: string[],
-        public email: string,
-        public firstName: string,
-        public lastName: string,
-        public login: string,
-        public imageUrl: string
-    ) { }
-}
