--- src/main/webapp/app/student-courses/student-courses.component.spec.ts
@@ -1,25 +1,0 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { StudentCoursesComponent } from './student-courses.component';
-
-describe('StudentCoursesComponent', () => {
-  let component: StudentCoursesComponent;
-  let fixture: ComponentFixture<StudentCoursesComponent>;
-
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [ StudentCoursesComponent ]
-    })
-    .compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(StudentCoursesComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should be created', () => {
-    expect(component).toBeTruthy();
-  });
-});
